Backend/requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.48
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
psycopg2==2.9.7
Flasgger==0.9.5
requests==2.31.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==2.3.7
gunicorn==21.2.0
Jinja2==3.1.2